Update to reflect go_highlight_function_calls
With the elimination of go_highlight_methods and the addition of
go_highlight_function_calls, the tutorial needs to be updated, as
does the sample included vimrc file.

As a newcomer to vim-go, it took me a few minutes to chase down
why I thought the syntax highlighting was not working (and almost
bailed back to vscode).